changed files

+++ install.sh
@@ -1,28 +1,36 @@
-#!/bin/sh
+#!/bin/zsh
+exec > >(tee ./tmp/install.log)
+exec 2>&1
+set -x
+set -e
 
 ## Execute permission 
-chmod +x archinstall1.sh
+chmod +x outside-arch-chroot.sh
 
 
 ## Before arch-chroot
-./archinstall1.sh 2>&1 | tee ./archinstall1.log
+./outside-arch-chroot.sh #2>&1 | tee ./tmp/archinstall1.log
 
 ## Place the archinstall2.sh script and grub.cfg
 ## at the root of the new filesystem (/mnt)
-cp ./archinstall2.sh /mnt/archinstall2.sh
-cp ./grub.cfg /mnt/grub.cfg
-cp ./archinstall1.log /mnt/archinstall1.log
+cp ./inside-arch-chroot.sh /mnt
+cp ./grub.cfg /mnt
 
 # Execute permission on archinstall2.sh
-chmod +x /mnt/archinstall2.sh
+chmod +x /mnt/inside-arch-chroot.sh
 
 
-## After arch-chroot
-arch-chroot /mnt /bin/sh -c "/archinstall2.sh 2>&1 | tee /archinstall2.log"
+## Inside chroot
+arch-chroot /mnt /bin/bash -c /inside-arch-chroot.sh
 
 ## remove the archinstall2.sh script from /mnt (the new root)
-rm -f /mnt/archinstall2.sh
+rm -f /mnt/inside-arch-chroot.sh
 
 # unmount and restart
 umount -R /mnt
-reboot
+
+read -q "key?Reboot [Yyn]? "
+if [[ "$key" =~ ^[Yy]$ ]]
+then
+    reboot
+fi
